refactor commands_impl to commands_shared_code
