+++ .gitignore
@@ -81,3 +81,5 @@ dist/
 # Node
 node_modules
 yarn-error.log
+
+.pre-release-scripts.txt
